Make iterateKeyEntityPairs reconstruct keys correctly or fail fast
BitmapIndex.iterateKeyEntityPairs inverts each stored long back to its index key via BinaryIndexer.binaryToKey. That inverse existed only on BinaryIndexerLong; the default returned the raw stored long. But every binary number indexer applies a non-identity encoding (unsigned mapping plus a zero sentinel, and for float/double a sortable-bit transform), so for those indexers the method silently emitted encoded/sentinel values instead of the original keys — e.g. int key 0 came out as 2^32.

Change the default binaryToKey to throw UnsupportedOperationException, so an indexer whose encoding cannot be inverted fails fast rather than yielding wrong keys, and add correct inverses for the integer indexers. Byte, Short and Integer encode a non-zero value as its unsigned representation and zero as 1L << SIZE; the inverse maps that sentinel back to 0 and re-narrows any other value with a signed cast ((byte)/(short)/(int)) to recover the original key. Float and Double keep the throwing default: their sortable-bit encoding is error-prone to invert and has no consumer today. binaryToKey is only called from iterateKeyEntityPairs, so the default change is contained; the sole production caller (VectorIndex, via BinaryIndexerLong) is unaffected.

Also finish the RandomAccessVectorValues.size() overflow hardening: the one remaining site, the test-only convenience constructor in GigaMapBackedVectorValues, now uses Math.toIntExact like the production size suppliers instead of a plain (int) cast.

Extend BitmapIndexKeyEntityPairsTest: reconstructsIntegerKeys_withSentinelAndNegatives checks an Integer identity index round-trips 0, +/-1, Integer.MIN_VALUE and a large value back to their original keys; floatIndex_failsFastAsUnsupported asserts a Float index throws UnsupportedOperationException rather than emitting encoded bits. Verified across the full gigamap core suite and the jvector fast and slow suites.

+++ gigamap/jvector/src/main/java/org/eclipse/store/gigamap/jvector/GigaMapBackedVectorValues.java
@@ -62,7 +62,7 @@ class GigaMapBackedVectorValues implements RandomAccessVectorValues
         final VectorTypeSupport    vectorTypeSupport
     )
     {
-        this(storeLookup(vectorStore), () -> (int)vectorStore.size(), dimension, vectorTypeSupport);
+        this(storeLookup(vectorStore), () -> Math.toIntExact(vectorStore.size()), dimension, vectorTypeSupport);
     }
 
     private static IntFunction<float[]> storeLookup(final GigaMap<VectorEntry> vectorStore)
